remove blog subfolder
