+++ src/routes/[slug]/+page.js
